requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
Flask==2.1.2
fonttools==4.33.3
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
imageio==2.19.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pydicom==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scipy==1.8.0
six==1.16.0
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
